Made some changes in file structure
