+++ app/src/main/java/com/aritra/medsync/navigation/MedSyncApp.kt
@@ -29,21 +29,21 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.aritra.medsync.domain.model.Medication
-import com.aritra.medsync.screens.addMedication.AddMedication
-import com.aritra.medsync.screens.addMedication.AddMedicationViewModel
-import com.aritra.medsync.screens.homeScreen.HomeScreen
-import com.aritra.medsync.screens.medicationConfirmation.MedicationConfirmationScreen
-import com.aritra.medsync.screens.SplashScreen
-import com.aritra.medsync.screens.appointment.AppointmentScreen
-import com.aritra.medsync.screens.history.HistoryScreen
-import com.aritra.medsync.screens.history.viewmodel.HistoryViewModel
-import com.aritra.medsync.screens.homeScreen.viewmodel.HomeViewModel
-import com.aritra.medsync.screens.medicationConfirmation.MedicationConfirmViewModel
-import com.aritra.medsync.screens.prescription.PrescriptionScreen
-import com.aritra.medsync.screens.profile.ProfileScreen
-import com.aritra.medsync.screens.report.ReportScreen
-import com.aritra.medsync.screens.settings.SettingsScreen
-import com.aritra.medsync.screens.settings.SettingsViewModel
+import com.aritra.medsync.ui.screens.addMedication.AddMedication
+import com.aritra.medsync.ui.screens.addMedication.AddMedicationViewModel
+import com.aritra.medsync.ui.screens.homeScreen.HomeScreen
+import com.aritra.medsync.ui.screens.medicationConfirmation.MedicationConfirmationScreen
+import com.aritra.medsync.ui.screens.SplashScreen
+import com.aritra.medsync.ui.screens.appointment.AppointmentScreen
+import com.aritra.medsync.ui.screens.history.HistoryScreen
+import com.aritra.medsync.ui.screens.history.viewmodel.HistoryViewModel
+import com.aritra.medsync.ui.screens.homeScreen.viewmodel.HomeViewModel
+import com.aritra.medsync.ui.screens.medicationConfirmation.MedicationConfirmViewModel
+import com.aritra.medsync.ui.screens.prescription.PrescriptionScreen
+import com.aritra.medsync.ui.screens.profile.ProfileScreen
+import com.aritra.medsync.ui.screens.report.ReportScreen
+import com.aritra.medsync.ui.screens.settings.SettingsScreen
+import com.aritra.medsync.ui.screens.settings.SettingsViewModel
 import com.aritra.medsync.ui.theme.Background
 import com.aritra.medsync.ui.theme.DMSansFontFamily
 import com.aritra.medsync.ui.theme.FadeIn
