--- app/src/main/java/com/aritra/medsync/screens/appointment/AppointmentScreen.kt
@@ -1,8 +1,0 @@
-package com.aritra.medsync.screens.appointment
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun AppointmentScreen() {
-
-}
