--- app/src/main/java/com/aritra/medsync/screens/homeScreen/usecase/FetchMedicationUseCase.kt
@@ -1,15 +1,0 @@
-package com.aritra.medsync.screens.homeScreen.usecase
-
-import com.aritra.medsync.domain.model.Medication
-import com.aritra.medsync.domain.repository.MedicationRepository
-import kotlinx.coroutines.flow.Flow
-import javax.inject.Inject
-
-class FetchMedicationUseCase @Inject constructor(
-    private val repository: MedicationRepository
-) {
-
-    fun getAllMedications(): Flow<List<Medication>> {
-        return repository.getAllMedications()
-    }
-}
