--- app/src/main/java/com/aritra/medsync/screens/prescription/PrescriptionScreen.kt
@@ -1,8 +1,0 @@
-package com.aritra.medsync.screens.prescription
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun PrescriptionScreen() {
-
-}
