--- app/src/main/java/com/aritra/medsync/screens/report/ReportScreen.kt
@@ -1,8 +1,0 @@
-package com.aritra.medsync.screens.report
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun ReportScreen() {
-
-}
