+++ app/src/main/java/com/aritra/medsync/ui/screens/appointment/AppointmentScreen.kt
@@ -1,0 +1,8 @@
+package com.aritra.medsync.ui.screens.appointment
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun AppointmentScreen() {
+
+}
