+++ app/src/main/java/com/aritra/medsync/ui/screens/homeScreen/usecase/FetchMedicationUseCase.kt
@@ -1,0 +1,15 @@
+package com.aritra.medsync.ui.screens.homeScreen.usecase
+
+import com.aritra.medsync.domain.model.Medication
+import com.aritra.medsync.domain.repository.MedicationRepository
+import kotlinx.coroutines.flow.Flow
+import javax.inject.Inject
+
+class FetchMedicationUseCase @Inject constructor(
+    private val repository: MedicationRepository
+) {
+
+    fun getAllMedications(): Flow<List<Medication>> {
+        return repository.getAllMedications()
+    }
+}
