+++ app/src/main/java/com/aritra/medsync/ui/screens/homeScreen/viewmodel/MedicationDetailsViewModel.kt
@@ -1,0 +1,18 @@
+package com.aritra.medsync.ui.screens.homeScreen.viewmodel
+
+import androidx.lifecycle.ViewModel
+import com.aritra.medsync.domain.extensions.runIO
+import com.aritra.medsync.domain.model.Medication
+import com.aritra.medsync.ui.screens.homeScreen.usecase.MedicationUpdateUseCase
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class MedicationDetailsViewModel @Inject constructor(
+    private val medicationUpdateUseCase: MedicationUpdateUseCase
+) : ViewModel() {
+
+    fun isMedicationTaken(medication: Medication, isMedicationTaken: Boolean) = runIO {
+        medicationUpdateUseCase.updateMedication(medication.copy(isTaken = isMedicationTaken))
+    }
+}
