+++ app/src/main/java/com/aritra/medsync/ui/screens/prescription/PrescriptionScreen.kt
@@ -1,0 +1,8 @@
+package com.aritra.medsync.ui.screens.prescription
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun PrescriptionScreen() {
+
+}
