+++ app/src/main/java/com/aritra/medsync/ui/screens/report/ReportScreen.kt
@@ -1,0 +1,8 @@
+package com.aritra.medsync.ui.screens.report
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun ReportScreen() {
+
+}
